property rename fixed
